+++ CHANGELOG.md
@@ -261,6 +261,7 @@
 - [9624](https://github.com/vegaprotocol/vega/issues/9624) - Fix get returns when window is larger than available returns
 - [9634](https://github.com/vegaprotocol/vega/issues/9634) - Do not allow to submit liquidity provision for pending LPs
 - [9636](https://github.com/vegaprotocol/vega/issues/9636) - include ersatz validator for validator ranking reward.
+- [9655](https://github.com/vegaprotocol/vega/issues/9655) - Make liquidity monitoring parameter required in market proposals validation
 
 ## 0.72.1